+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadTrapDoorBlock.java
@@ -1,0 +1,145 @@
+package net.geforcemods.securitycraft.blocks;
+
+import java.util.Random;
+
+import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.api.INameSetter;
+import net.geforcemods.securitycraft.api.IOwnable;
+import net.geforcemods.securitycraft.api.IPasscodeConvertible;
+import net.geforcemods.securitycraft.api.IPasscodeProtected;
+import net.geforcemods.securitycraft.blockentities.KeypadTrapdoorBlockEntity;
+import net.geforcemods.securitycraft.blocks.reinforced.BaseIronTrapDoorBlock;
+import net.geforcemods.securitycraft.misc.OwnershipEvent;
+import net.geforcemods.securitycraft.misc.SaltData;
+import net.geforcemods.securitycraft.util.PlayerUtils;
+import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.state.properties.Half;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.ActionResultType;
+import net.minecraft.util.Direction;
+import net.minecraft.util.Hand;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.BlockRayTraceResult;
+import net.minecraft.util.text.TextFormatting;
+import net.minecraft.world.IBlockReader;
+import net.minecraft.world.World;
+import net.minecraft.world.server.ServerWorld;
+import net.minecraftforge.common.MinecraftForge;
+
+public class KeypadTrapDoorBlock extends BaseIronTrapDoorBlock {
+	public KeypadTrapDoorBlock(Properties properties) {
+		super(properties);
+	}
+
+	@Override
+	public ActionResultType use(BlockState state, World level, BlockPos pos, PlayerEntity player, Hand hand, BlockRayTraceResult hit) {
+		if (state.getValue(OPEN))
+			return ActionResultType.PASS;
+		else if (!level.isClientSide) {
+			KeypadTrapdoorBlockEntity be = (KeypadTrapdoorBlockEntity) level.getBlockEntity(pos);
+
+			if (be.isDisabled())
+				player.displayClientMessage(Utils.localize("gui.securitycraft:scManual.disabled"), true);
+			else if (be.verifyPasscodeSet(level, pos, be, player)) {
+				if (be.isDenied(player)) {
+					if (be.sendsMessages())
+						PlayerUtils.sendMessageToPlayer(player, Utils.localize(getDescriptionId()), Utils.localize("messages.securitycraft:module.onDenylist"), TextFormatting.RED);
+				}
+				else if (be.isAllowed(player)) {
+					if (be.sendsMessages())
+						PlayerUtils.sendMessageToPlayer(player, Utils.localize(getDescriptionId()), Utils.localize("messages.securitycraft:module.onAllowlist"), TextFormatting.GREEN);
+
+					activate(state, level, pos, be.getSignalLength());
+				}
+				else if (!PlayerUtils.isHoldingItem(player, SCContent.CODEBREAKER, hand))
+					be.openPasscodeGUI(level, pos, player);
+			}
+		}
+
+		return ActionResultType.SUCCESS;
+	}
+
+	public void activate(BlockState state, World level, BlockPos pos, int signalLength) {
+		level.setBlockAndUpdate(pos, state.setValue(OPEN, true));
+		level.getBlockTicks().scheduleTick(pos, this, signalLength);
+		playSound(null, level, pos, true);
+	}
+
+	@Override
+	public void tick(BlockState state, ServerWorld level, BlockPos pos, Random random) {
+		if (state.getValue(OPEN)) {
+			level.setBlockAndUpdate(pos, state.setValue(OPEN, false));
+			playSound(null, level, pos, false);
+		}
+	}
+
+	//here for making it accessible without AT
+	@Override
+	public void playSound(PlayerEntity player, World level, BlockPos pos, boolean isOpened) {
+		super.playSound(player, level, pos, isOpened);
+	}
+
+	@Override
+	public void onRemove(BlockState state, World level, BlockPos pos, BlockState newState, boolean isMoving) {
+		if (!state.is(newState.getBlock())) {
+			TileEntity be = level.getBlockEntity(pos);
+
+			if (be instanceof IPasscodeProtected)
+				SaltData.removeSalt(((IPasscodeProtected) be).getSaltKey());
+
+			super.onRemove(state, level, pos, newState, isMoving);
+		}
+	}
+
+	@Override
+	public void setPlacedBy(World level, BlockPos pos, BlockState state, LivingEntity placer, ItemStack stack) {
+		if (placer instanceof PlayerEntity)
+			MinecraftForge.EVENT_BUS.post(new OwnershipEvent(level, pos, (PlayerEntity) placer));
+
+		if (stack.hasCustomHoverName()) {
+			TileEntity be = level.getBlockEntity(pos);
+
+			if (be instanceof INameSetter)
+				((INameSetter) be).setCustomName(stack.getHoverName());
+		}
+	}
+
+	@Override
+	public void neighborChanged(BlockState state, World level, BlockPos pos, Block block, BlockPos neighbor, boolean flag) {}
+
+	@Override
+	public TileEntity createTileEntity(BlockState state, IBlockReader world) {
+		return new KeypadTrapdoorBlockEntity();
+	}
+
+	public static class Convertible implements IPasscodeConvertible {
+		@Override
+		public boolean isValidStateForConversion(BlockState state) {
+			return state.is(SCContent.REINFORCED_IRON_TRAPDOOR.get());
+		}
+
+		@Override
+		public boolean convert(PlayerEntity player, World level, BlockPos pos) {
+			BlockState state = level.getBlockState(pos);
+			Direction facing = state.getValue(FACING);
+			boolean open = state.getValue(OPEN);
+			Half half = state.getValue(HALF);
+			boolean powered = state.getValue(POWERED);
+			boolean waterlogged = state.getValue(WATERLOGGED);
+			TileEntity trapdoor = level.getBlockEntity(pos);
+			CompoundNBT tag = trapdoor.save(new CompoundNBT());
+
+			level.setBlockAndUpdate(pos, SCContent.KEYPAD_TRAPDOOR.get().defaultBlockState().setValue(FACING, facing).setValue(OPEN, open).setValue(HALF, half).setValue(POWERED, powered).setValue(WATERLOGGED, waterlogged));
+			trapdoor = level.getBlockEntity(pos);
+			trapdoor.load(null, tag);
+			((IOwnable) trapdoor).setOwner(player.getUUID().toString(), player.getName().getString());
+			return true;
+		}
+	}
+}