+++ Changelog.md
@@ -8,6 +8,7 @@
 - New: Maximum Distance option for Retinal Scanner and Scanner Door, to define the maximum amount of blocks from which the block can be activated
 - New: Passcodes entered in the text box of check passcode screens can now be shown using a checkbox
 - New: Keycard Holder
+- New: Keypad Trapdoor
 - Change: Several technical blocks' sounds have been adjusted to better match how they look
 - Change: Improved visuals when holding a taser
 - Change: Some alarm options have been moved to a separate screen